NOVAS
The Naval Observatory Vector Astrometry Software (NOVAS) library is used to accurately calculate the positions of astronomical objects, such as stars and planets. It is the basis for the annual Astronomical Almanac produced by the U.S. Naval Observatory and the Her Majesty's Nautical Almanac Office. We propose developing a Julia version of this software and improving its performance by including vector calculations, which is critical for big data applications.